Create KIC rejection application
This Perl script - accompanied by its deployment tool - is the source file of a stand-alone application, which helps document validation operators reject problematic Accounts Payable (AP) documents in an invoice validation workflow in Kofax. The main purpose of the application is to cross-reference and identify active Kofax batches and their associated Kofax Import Connector source messages in an automated way and to generate rejection emails to the AP team indicating the rejection reason. The only front-end interface for the application is an executable utility, therefore - although, it's written entirely in Perl - it's self-contained, there's no need to install the Perl interpreter or any prerequisite modules to run it.